Some work on Linux compatibility
Modified tweet footer, because of wx errors on this platform

Multiline text controls should not have a maximum length

Replaced stream.py from sound_lib with another found on my disk. So, sounds should play on Linux

You should replace it with this located in this repository.

Modified sizes of some controls in visible interface to avoid gtk errors